refactor(misc1): move shell related functions to os/shell.c
